refactor: move and simplify tx building and simulation
